+++ src/Notes.tsx
@@ -1,31 +1,23 @@
-import { useStore } from '@tanstack/react-store';
-import { musicSheet } from './music-sheet';
 import './Notes.css'
 
-function Notes() {
-    const displayNoteCount = 10;
-    const first: number[] = useStore(musicSheet, (state) => state.notes.first(displayNoteCount));
+export type NotesProps = {notes: number[], count:number};
 
-    function notes() {
-        const rows = [];
-        for (let i = 0; i < displayNoteCount; i++) {
-            if(first.at(i) === undefined) {
-                rows.push(<span className='number-note' key={i}>_</span>);
-            } else {
-                rows.push(<span className='number-note' key={i}>{first.at(i)}</span>);                
-            }
+export function Notes({notes, count}: NotesProps) {
+    const rows = [];
+    for (let i = 0; i < count; i++) {
+        if(notes.at(i) === undefined) {
+            rows.push(<span className='number-note' key={i}>_</span>);
+        } else {
+            rows.push(<span className='number-note' key={i}>{notes.at(i)}</span>);                
         }
-        return rows
     }
+
     return (
-        <>
-            <div className='container'>
-                <div className='notes'>
-                    { notes() }
-                </div>
+        <div className='container'>
+            <div className='notes'>
+                { rows }
             </div>
-        </>
+        </div>
     )
 }
 
-export default Notes
