--- src/base.ts
@@ -1,7 +1,0 @@
-import { Store } from "@tanstack/store";
-
-
-// You can use @tanstack/store outside of React components too!
-export const base = new Store({
-  currentBase: 7
-});
