--- src/music-scales.ts
@@ -1,12 +1,0 @@
-export const C_MAJOR_PENTATONIC = [
-    "C3",  // index 0
-    "D3",  // index 1
-    "E3",  // index 2
-    "G3",  // index 3
-    "A3",  // index 4
-    "C4",  // index 5
-    "D4",  // index 6
-    "E4",  // index 7
-    "G4",  // index 8
-    "A4"   // index 9
-];
